The SOC decomposition model

This function implements with the SoilR model development framework the dual porosity model described in Meurer et al. (2020). The model is an evolution of a two-pool linear SOC model, with two pools (young and old material) running in parallel for micro- and mesopores. While aboveground inputs are rooted in the mesopores, root inputs are distributed between micro and mesopores depending on porosity, which is in turn influenced by organic matter. This makes the model nonlinear, although it still behaves similarly to a linear model within a reasonable calibration range. The model is described by a series of four equations:

\(\frac{dM_{Y_{(mes)}}}{dt} = I_m + \left( \frac{\phi_{mes}}{\phi_{mes}+\phi_{mic}}\right) \cdot I_r - k_Y \cdot M_{Y_{(mes)}}+ T_Y \)
\( \frac{dM_{O_{(mes)}}}{dt} = \left( \epsilon \cdot k_Y \cdot M_{Y_{(mes)}} \right) - \left( (1- \epsilon) \cdot k_O \cdot M_{O_{(mes)}} \right) + T_O\)
\( \frac{dM_{Y_{(mic)}}}{dt} = \left( \frac{\phi_{mic}}{\phi_{mes}+\phi_{mic}}\right) \cdot I_r - k_Y \cdot F_{prot} \cdot M_{Y_{(mes)}}- T_Y \)
\( \frac{dM_{O_{(mic)}}}{dt} = \left( \epsilon \cdot k_Y \cdot F_{prot} \cdot M_{Y_{(mes)}} \right) - \left( (1- \epsilon) \cdot k_O \cdot F_{prot} \cdot M_{O_{(mes)}} \right) - T_O \)

Please refer to the original paper for more details.
The terms \(T_Y\) and \(T_Y\) are calculated in the original paper as \(k_{mix} \cdot \frac{(My_{mic}-My_{mes})}{2}\) and \(k_{mix} \cdot \frac{(Mo_{mic}-Mo_{mes})}{2}\), but in a later model development the term 2 disappeared and are now calculaged as \(T_Y = k_{mix} \cdot (My_{mic}-My_{mes})\) and \(T_O = k_{mix} \cdot (Mo_{mic}-Mo_{mes})\).

The two porosity terms, \(\phi_{mes} = f(M_{Y_{(mes)}}, M_{O_{(mes)}},M_{Y_{(mic)}}, M_{O_{(mic)}})\) and \(\phi_{mic} = f(M_{Y_{(mic)}}, M_{O_{(mic)}})\), are dependent on the variation of the different C pools and everything is variable over time, introducing a nonlinearity in the system and defining the biggest peculiarity of this model.
After substituting the terms \(\left( \frac{\phi_{mes}(t)}{\phi_{mes}(t)+\phi_{mic}(t)}\right) = \varphi_{mes}\) and \(\left( \frac{\phi_{mic}(t)}{\phi_{mes}(t)+\phi_{mic}(t)}\right) = \varphi_{mic}\), The model can be rewritten in matrix form as :

\(I_m(t) + I_r(t) \cdot N(C,t) + A(t) \cdot P(t) \cdot C(t)\)
The model is implemented with the SoilR package, but it is relying on a more conventional ODE definition (not its matrix form).

Porous(
  ky,
  ko,
  kmix,
  e,
  Im,
  Ir,
  F_prot,
  phi_mac,
  phi_min,
  clay,
  Delta_z_min,
  gamma_o,
  proportion = NULL,
  f_text_mic = NULL,
  f_agg
)

Arguments

ky

decomposition constant of the Young pool \(frac{1}{year}\)

ko

decomposition constant of the Old pool \(frac{1}{year}\)

kmix

mixing rate \(frac{1}{year}\)

e

efficiency, which is the transfer term between the pools and corresponds to the term h in the ICBM model in Kätterer et al. (2001) (dimensionless)

F_prot

protection provided by the micropore space (dimensionless)

phi_mac

macroporosity, (\(\frac{cm^3 of water}{cm^3 of soil}\))

phi_min

minimal porosity, (\(\frac{cm^3 of water}{cm^3 of soil}\))

clay

fraction of clay content (dimensionless, between 0 and 1)

Delta_z_min

minimal soil thickness if no organic matter was present (cm)

proportion

this is a linearization term to make the proportion of the inputs between micro- and mesopores constant. If NULL (or not specified, since default is NULL) then the model is running as nonlinear, as in the original paper. If specified (must be between 0 and 1) then the model is linearized adopting this value as fixed proportion of inputs from roots going into the mesopore space (and its reciprocal into the micropore)

f_agg

an aggregation factor (m3 pore space m-3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \(V_{agg}\), and the volume of organic matter \(V_{s_o}\) (dimensionless)

Value

two values, the proportion (between 0 and 1) of input in the mesopore and micropore Y pools

References

Meurer, Katharina Hildegard Elisabeth, Claire Chenu, Elsa Coucheney, Anke Marianne Herrmann, Thomas Keller, Thomas Kätterer, David Nimblad Svensson, and Nicholas Jarvis. “Modelling Dynamic Interactions between Soil Structure and the Storage and Turnover of Soil Organic Matter.” Biogeosciences 17, no. 20 (October 19, 2020): 5025–42. https://doi.org/10.5194/bg-17-5025-2020.
Kätterer, Thomas, and Olof Andrén. “The ICBM Family of Analytically Solved Models of Soil Carbon, Nitrogen and Microbial Biomass Dynamics — Descriptions and Application Examples.” Ecological Modelling 136, no. 2–3 (January 2001): 191–207. https://doi.org/10.1016/S0304-3800(00)00420-8.

See also