Run Porous

This function is a wrapper for running the main model Porous, which is a wrapper for the dual porosity decomposition model (equations 1 to 6) described in Meurer et al. (2020) implemented in `SoilR`. It then feeds the simulated SOC stocks to the functions gamma_b and f_som to calculate the variation of bulk density and C concentrations

run_Porous(
  ky = 0.8,
  ko = 0.04,
  kmix = 0.02,
  e = 0.15,
  Im = 0.08,
  Ir = 0.048,
  F_prot = 0.1,
  phi_mac = 0.04,
  phi_min = 0.35,
  clay = 0.2,
  Delta_z_min = 14.2,
  gamma_o = 1.2,
  gamma_m = 2.7,
  proportion = NULL,
  f_text_mic = NULL,
  f_agg = 3,
  init,
  sim_length,
  sim_steps
)

Arguments

ky

decomposition constant of the Young pool \(frac{1}{year}\)

ko

decomposition constant of the Old pool \(frac{1}{year}\)

kmix

mixing rate \(frac{1}{year}\)

e

efficiency, which is the transfer term between the pools and corresponds to the term h in the ICBM model in Kätterer et al. (2001) (dimensionless)

F_prot

protection provided by the micropore space (dimensionless)

phi_mac

macroporosity, (\(\frac{cm^3 of water}{cm^3 of soil}\))

phi_min

minimal porosity, (\(\frac{cm^3 of water}{cm^3 of soil}\))

clay

fraction of clay content (dimensionless, between 0 and 1)

Delta_z_min

minimal soil thickness if no organic matter was present (cm)

proportion

this is a linearization term to make the proportion of the inputs between micro- and mesopores constant. If NULL (or not specified, since default is NULL) then the model is running as nonlinear, as in the original paper. If specified (must be between 0 and 1) then the model is linearized adopting this value as fixed proportion of inputs from roots going into the mesopore space (and its reciprocal into the micropore)

f_agg

an aggregation factor (m3 pore space m-3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \(V_{agg}\), and the volume of organic matter \(V_{s_o}\) (dimensionless)

sim_steps

steps of the simulation (fraction of one year)

sim_lenght

length of the simulation (years)

Value

a data frame with the simulation of C stocks for each of the four pools, soil respiration for each of the four pools, bulk density variation and SOC concentration.

See also