The SOC decomposition model

This function implements with the SoilR model development framework the dual porosity model described in Meurer et al. (2020). The model is an evolution of a two-pool linear SOC model, with two pools (young and old material9) running in parallel for micro and mesopores. While aboveground inputs are rooted in the mesopores, root inputs are distributed between micro and mesopores depending on porosity, which is in turn influenced by organic matter. This makes the model nonlinear, although it still behaves similarly to a linear model within a reasonable calibration range. The model is described by a series of four equations:

\(\frac{dM_{Y_{(mes)}}}{dt} = I_m + \left( \frac{\phi_{mes}}{\phi_{mes}+\phi_{mic}}\right) \cdot I_r - k_Y \cdot M_{Y_{(mes)}}+ T_Y \)
\( \frac{dM_{O_{(mes)}}}{dt} = \left( \epsilon \cdot k_Y \cdot M_{Y_{(mes)}} \right) - \left( (1- \epsilon) \cdot k_O \cdot M_{O_{(mes)}} \right) + T_O\)
\( \frac{dM_{Y_{(mic)}}}{dt} = \left( \frac{\phi_{mic}}{\phi_{mes}+\phi_{mic}}\right) \cdot I_r - k_Y \cdot F_{prot} \cdot M_{Y_{(mes)}}- T_Y \)
\( \frac{dM_{O_{(mic)}}}{dt} = \left( \epsilon \cdot k_Y \cdot F_{prot} \cdot M_{Y_{(mes)}} \right) - \left( (1- \epsilon) \cdot k_O \cdot F_{prot} \cdot M_{O_{(mes)}} \right) - T_O \)

Please refer to the original paper for more details.

The two porosity terms, \(\phi_{mes} = f(M_{Y_{(mes)}}, M_{O_{(mes)}},M_{Y_{(mic)}}, M_{O_{(mic)}})\) and \(\phi_{mic} = f(M_{Y_{(mic)}}, M_{O_{(mic)}})\), are dependent on the variation of the different C pools and everything is variable over time, introducing a nonlinearity in the system and defining the biggest peculiarity of this model.
After substitutung the terms \(\left( \frac{\phi_{mes}(t)}{\phi_{mes}(t)+\phi_{mic}(t)}\right) = \varphi_{mes}\) and \(\left( \frac{\phi_{mic}(t)}{\phi_{mes}(t)+\phi_{mic}(t)}\right) = \varphi_{mic}\), The model can be rewritten in matrix form as :

\(I_m(t) + I_r(t) \cdot N(C,t) + A(t) \cdot P(t) \cdot C(t)\)
Or, more explicitly:

\(\frac{dC}{dt}=\begin{bmatrix} I_m\\ 0\\ 0\\ 0\\ \end{bmatrix}+ \begin{bmatrix} I_r\\ 0\\ I_r\\ 0\\ \end{bmatrix} \cdot \begin{bmatrix} \varphi_{mes} & 0 & 0 & 0\\ 0 & 1 & 0 & 0\\ 0 & 0 & \varphi_{mic} & 0\\ 0 & 0 & 0 & 1\\ \end{bmatrix}+ \begin{bmatrix} -k_y & \epsilon & 0 & 0\\ 0 & -k_o & 0 & 0\\ T_Y & 0 & -k_y & \epsilon\\ 0 & T_O & 0 & -k_o\\ \end{bmatrix} \cdot \begin{bmatrix} 1 & 0 & 0 & 0\\ 0 & 1 & 0 & 0\\ 0 & 0 & F_{prot} & 0\\ 0 & 0 & 0 & F_{prot}\\ \end{bmatrix} \cdot \begin{bmatrix} M_{Y_{mes}}\\ M_{O_{mes}}\\ M_{Y_{mic}}\\ M_{O_{mic}} \end{bmatrix} \)

Porous(
  ky = 0.8,
  ko = 0.00605,
  kmix = 0.9,
  e = 0.13,
  Im = 1.1,
  Ir = 0.5,
  F_prot = 0,
  phi_mac = 0.2,
  clay = 0.2,
  Delta_z_min = 20,
  gamma_o = 1.2
)

Arguments

ky

decomposition constant of the Young pool

ko

decomposition constant of the Old pool

kmix

mixing rate

e

efficiency, which is the transfer term between the pools and corresponds to the term h in the ICBM model in Kätterer et al. (2001)

Im

Inputs from aboveground

Ir

Inputs from roots

F_prot

protection provided by the micropore space

phi_mac

macroporosity

clay

fraction of clay content

Delta_z_min

minimal soil thickness if no organic matter was present

gamma_o

density of organic matter

Value

two values, the proportion of input in the mesopore and micropore Y pools

References

Meurer, Katharina Hildegard Elisabeth, Claire Chenu, Elsa Coucheney, Anke Marianne Herrmann, Thomas Keller, Thomas Kätterer, David Nimblad Svensson, and Nicholas Jarvis. “Modelling Dynamic Interactions between Soil Structure and the Storage and Turnover of Soil Organic Matter.” Biogeosciences 17, no. 20 (October 19, 2020): 5025–42. https://doi.org/10.5194/bg-17-5025-2020.
Kätterer, Thomas, and Olof Andrén. “The ICBM Family of Analytically Solved Models of Soil Carbon, Nitrogen and Microbial Biomass Dynamics — Descriptions and Application Examples.” Ecological Modelling 136, no. 2–3 (January 2001): 191–207. https://doi.org/10.1016/S0304-3800(00)00420-8.